ensure uniform post-trailer-zeros length
After the trailer entry, copy-out currently writes out zeros up to the
next block boundary. When reflink is enabled, writes may be unbuffered.
To ensure that the post-trailer-zeros length matches with and without
reflink_flag, calculate the zeros length based on where the I/O buffer
offset would be instead of where it currently is.

Signed-off-by: David Disseldorp <ddiss@suse.de>